Fixed cookie expiry bug; added autofinger list method
plans_login now returns True or False to indicate whether our login
succeeded, instead of raising an exception. You can use it with no
arguments to determine whether a login is active or not.

This enables us to prompt for password if our cookie is expired.

I also started adding a method for the autofinger list, since there's an
easy JSON interface existing on the server. No functionality for that
yet, but it seems to sort of work.