Strict unicode handling, strict MD5 hashing
This is a risky move, but I added explicit encodes and decodes such
that clans is handling unicode objects wherever possible (instead of
strings). This "unicode sandwich" approach should help prevent WTF
bugs in the future.

Also, it appears that I've figured out the MD5 hashing. When editing a
plan, clans now verifies the MD5 hash computed on the server. SWEET.

ADDED BONUS: if you save a backup copy of a plan (with clans --backup
filename.plan --skip-update) and run an md5 on the generated file, it
will match the one computed on the server. DOUBLE SWEET.