set_edit_text scrapes for warnings now
Doesn't act on any of these warnings, just displays them. Eventually
we will want to have a sensible UI for returning to the edit interface
after errors are thrown, so that people don't lose data.

Added a test for really long plans.

for really REALLY long plans, the alertmessage is not generated, for
some reason.