Added (failing) test for read output
I've decided that the read_plan method should not do excessive
processing - it should simply return the plan text as it exists in the
database. This isn't so simple, because I am scraping to get at it.
Might as well have a failing test, and get as close as I can.