special-cased config command
This fixes the annoyance of having to log in to edit a local
configuration file.

I don't like the way I have implemented this. It should really be up
to the subcommand whether they want an active plans login or not.

The solution is to spin the cookie-baking, logging-in section of
main() into a function that can be called from each subcommand that
needs it. It would be cool to do this as a function decorator but I
don't have the energy to implement that right now