Remove arbitrary .py files as extensions
This was shoddily done, and I never used or tested it. I think it
would have precluded implementing extensions more generally (as a
folder with __init__.py, for example).

Better to get on board the python packaging train, and implement clans
extensions outside the standard set as their own python modules which
you install with pip.