Remove with_formatter decorator
That was a dumb idea, what's the point of putting the formatter object
in the function argument if the decorator is just going to consume it.

Explicit is better: in the function argument, but optional - if None,
call a method on the clans session which knows how to make the right
one. Since it remains in the argument list it is possible to test
subcommand functions by passing in a fake formatter (dependency
injection).

Will do this for the plans connection object shortly.