Explicitly handle line endings
Edit text should use CRLF endings, so that when transmitted as form
data it conforms to the W3C spec. This is also important for hashing:
Plans stores and md5-hashes the edit text with CRLF endings, so using
this ending throughout makes it easy for us to verify on the client
side as well.

See:
http://blogs.warwick.ac.uk/chrismay/entry/how_are_newlines/
http://www.w3.org/TR/REC-html40/interact/forms.html#h-17.13.4

Enforcing line endings on html plans is a more opinionated choice.
Reputable parsers such as html5lib basically convert everything to
'\n' endings before they even tokenize the thing, as dictated by the
standard:

http://www.whatwg.org/specs/web-apps/current-work/multipage/parsing.html#preprocessing-the-input-stream

So I'll do the same, even though what plans currently returns uses
'\r' line endings. That behavior is something I consider a bug, since
outside the plan there are exclusively '\n' endings.

This will also make the output of 'clans read' use consistent line
endings, and make it straightforward to upgrade to html5lib as our
parser.

This commit is a modified version of the changes made to the master
branch.