change ClansSession.args to dict (not namespace)
I find argparse's Namespace to be a little frivolous - the native data
structure to use is a dictionary, so let's use that.

Bug fixed: enabling the newlove extension, but not configuring it to
stalk, previously made `clans search` crash. Now that cs.args is a
dictionary, we can elegantly tiptoe around nonexistent arguments using
the get() method of dictionaries.