Get integration tests passing on py33
This was actually a nightmare, because it's difficult to deal with
stdout in a py3/py2 agnostic way (I really don't want to use
os.write(1, bytestring) since that feels so janky). I did a try/except
block which isn't so bad.

Getting tests to run properly was even more of a cluster due to a
combination of the above and bytes/unicode differences in temporary
files. I got it to work, but only with a mess of .decode('utf8')
method calls everywhere.

Basically, what makes single-codebase py2/py3 difficult is that
certain APIs that took bytestrings before take unicode strings now. In
the case of open files, it's easy to just do things the python 3 way
everywhere, since py26 and py27 have the open() function backported as
io.open(). However, for APIs that open files for you (tempfile
constructors, stdout, etc) you have no control over what open()
function they use internally, and so you have to do things the python
2 way.

Also, the integration tests are just dog slow, and testing the backup
extension by launching a subprocess often obscures the error output.
Really need unit test coverage of this stuff sometime, but for now I
really just want to land this branch.