renamed --file to --from-file, doc changes
--file is a dangerous option, so I removed the short flag -f and
renamed the long flag to something more explicit. Ideally this option
should be part of the backup extension but that's not possible right
now because the hooks don't offer any way to modify the plan text.

I also made a lot of documentation additions/changes that have been
nagging me for a while, and changed a few of the command line --help
outputs to flow more nicely together.