Make HTTPS the default config value.
I thought HTTPS was being used by default as of a few months ago, but
the config default specified in ui.py overrides the default argument
specified in scraper.py, and I forgot to change that one. Fixed.

Thanks Alex Cohn for making HTTPS mandatory on the server side and
thereby revealing this bug to me